refactor(responsive): module responsive
Remove service from module support and add it's own module.
