+++ libs/responsive/package.json
@@ -1,8 +1,33 @@
 {
   "name": "@ng-lab/responsive",
-  "version": "0.0.1",
+  "version": "0.0.0-PLACEHOLDER",
   "peerDependencies": {
     "@angular/common": "^7.1.0",
-    "@angular/core": "^7.1.0"
-  }
+    "@angular/core": "^7.1.0",
+    "@ng-lab/support": "0.0.0-PLACEHOLDER"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "private": false,
+  "author": {
+		"name": "Miguel Ramos",
+		"email": "miguel.marques.ramos@gmail.com",
+		"url": "https://ng-lab.github.com/"
+	},
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/miguelramos/ng-lab"
+	},
+	"homepage": "https://github.com/miguelramos/ng-lab",
+	"bugs": {
+		"url": "https://github.com/miguelramos/ng-lab/issues"
+	},
+	"keywords": [
+		"angular",
+    "library",
+    "ng-lab",
+    "responsive"
+	]
 }
