+++ libs/responsive/src/lib/responsive.module.spec.ts
@@ -1,14 +1,21 @@
 import { async, TestBed } from '@angular/core/testing';
 import { ResponsiveModule } from './responsive.module';
+import { ResponsiveBreakpointDirective } from './responsive.directive';
 
-describe('ResponsiveModule', () => {
+describe('> ResponsiveModule', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [ResponsiveModule]
     }).compileComponents();
   }));
 
-  it('should create', () => {
+  it('# Should have responsible module defined.', () => {
     expect(ResponsiveModule).toBeDefined();
+  });
+
+  it('# Should have ResponsiveBreakpointDirective on dependency injector.', () => {
+    const service = TestBed.get(ResponsiveBreakpointDirective, null);
+
+    expect(typeof service === 'object').toBeTruthy();
   });
 });
