+++ libs/support/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/support';
 export * from './lib/hub.service';
 export * from './lib/support.module';
-export { RESPONSIVE_BREAKPOINTS_TOKEN, ResponsiveBreakpointDirective } from './lib/responsive.directive';
