+++ libs/support/src/lib/support.module.spec.ts
@@ -2,7 +2,6 @@ import { async, TestBed } from '@angular/core/testing';
 
 import { HubService } from './hub.service';
 import { SupportModule } from './support.module';
-import { ResponsiveBreakpointDirective } from './responsive.directive';
 
 describe('> SupportModule', () => {
   beforeEach(async(() => {
@@ -21,9 +20,5 @@ describe('> SupportModule', () => {
     expect(typeof service === 'object').toBeTruthy();
   });
 
-  it('# Should have ResponsiveBreakpointDirective on dependency injector.', () => {
-    const service = TestBed.get(ResponsiveBreakpointDirective, null);
 
-    expect(typeof service === 'object').toBeTruthy();
-  });
 });
