+++ package.json
@@ -24,9 +24,10 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "npm run build:support && npm run build:configurator",
+    "build": "npm run build:support && npm run build:configurator && npm run build:responsive",
     "build:support": "ng build support",
     "build:configurator": "ng build configurator",
+    "build:responsive": "ng build responsive",
     "test": "ng test",
     "lint": "./node_modules/.bin/nx lint && ng lint",
     "e2e": "ng e2e",
